requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==3.0.1
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
easyocr==1.7.1
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
h11==0.14.0
idna==3.6
imageio==2.33.1
Jinja2==3.1.3
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
opencv-python-headless==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
pyclipper==1.3.0.post5
pycparser==2.21
PySocks==1.7.1
python-bidi==0.4.2
pytz==2023.3.post1
PyYAML==6.0.1
qq-bot==0.8.5
qq-botpy==1.2.0
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
selenium==4.16.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
tifffile==2023.12.9
torch==2.1.2
torchvision==0.16.2
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
websocket-client==1.7.0
wsproto==1.2.0
yarl==1.9.4
